Add InteractionManager prefab and update scene references
Created a new InteractionManager prefab under Managers and updated the Nandan scene to use this prefab instance instead of an in-scene object. Moved InventoryManager prefab from UI to Managers. Added new serialized fields to interactable MonoBehaviours in the scene for time and interaction management. Updated NavMesh asset reference in the scene.